+++ src/types/index.d.ts
@@ -12,6 +12,8 @@ declare global {
 
 export type Token = jwt.Options;
 
+export type Factory<Entity> = (data?: Entity) => Promise<Entity> | Entity;
+
 export interface IUserInputDTO {
   firstName: string;
   lastName: string;
@@ -22,17 +24,4 @@ export interface IUserInputDTO {
 export interface IUserResponseDTO {
   user: User;
   token: string;
-}
-
-export interface ICompanyInputDTO {
-  name: string;
-  description: string;
-  logo: string;
-  website: string;
-  headquarters: {
-    city: string;
-    country: string;
-  };
-  industry: string;
-  foundedYear: string;
 }