+++ test/integration/companyService.spec.ts
@@ -1,22 +1,27 @@
 import { Container } from 'typedi';
 import CompanyService from '../../src/api/services/company';
 import databaseLoader from '../../src/loaders/database';
-import CompanySeed from '../../src/database/seeds/CompanySeed';
 import { Connection } from 'typeorm';
 import Logger from '../../src/logger';
 import CompanyFactory from '../../src/database/factories/CompanyFactory';
 import { Company } from '../../src/api/entities/Company';
+import EntitySeed from '../../src/database/seeds/EntitySeed';
 jest.mock('../../src/logger');
 
 describe('CompanyService', () => {
   let connection: Connection;
-  let companySeed: CompanySeed;
+  let companySeed: EntitySeed<Company>;
+  let companyServiceInstance: CompanyService;
   beforeAll(async (done) => {
     Container.reset();
     connection = await databaseLoader();
     await connection.synchronize(true);
-    companySeed = new CompanySeed(connection);
+    companySeed = new EntitySeed<Company>(
+      connection.getMongoRepository(Company),
+      CompanyFactory
+    );
     Container.set('logger', Logger);
+    companyServiceInstance = Container.get(CompanyService);
     done();
   });
 
@@ -35,7 +40,6 @@ describe('CompanyService', () => {
   describe('create', () => {
     test('Should successfully create a company record', async () => {
       const mockCompany = CompanyFactory();
-      const companyServiceInstance = Container.get(CompanyService);
       const response = await companyServiceInstance.create(mockCompany);
 
       expect(response).toBeDefined();
@@ -45,7 +49,6 @@ describe('CompanyService', () => {
 
     test('Should fail to create a company record if the company name already exists', async () => {
       const existingCompany = await companySeed.seedOne();
-      const companyServiceInstance = Container.get(CompanyService);
       let err: Error, response: Company;
       try {
         response = await companyServiceInstance.create(existingCompany);
@@ -53,39 +56,7 @@ describe('CompanyService', () => {
         err = e;
       }
       expect(response).toBeUndefined();
-      expect(err).toEqual(new Error('The company already exists'));
-    });
-  });
-
-  describe('find', () => {
-    test('Should find all the companies', async () => {
-      const mockCompanies = await companySeed.seedMany(5);
-      const companyServiceInstance = Container.get(CompanyService);
-      const response = await companyServiceInstance.find();
-
-      expect(response).toBeDefined();
-      expect(response.sort()).toEqual(mockCompanies.sort());
-    });
-  });
-
-  describe('findOne', () => {
-    test('Should find a company with the valid id', async () => {
-      const mockCompanies = await companySeed.seedMany(5);
-      const companyServiceInstance = Container.get(CompanyService);
-      const response = await companyServiceInstance.findOne(
-        mockCompanies[0].id.toHexString()
-      );
-
-      expect(response).toBeDefined();
-      expect(response).toEqual(mockCompanies[0]);
-    });
-
-    test('Should return an error with an invalid id', async () => {
-      const mockCompanyId = '22dba00215a1568fe9310409';
-      const companyServiceInstance = Container.get(CompanyService);
-      const response = await companyServiceInstance.findOne(mockCompanyId);
-
-      expect(response).toBeUndefined();
+      expect(err).toEqual(new Error('The Company already exists'));
     });
   });
 });